% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{compute_richness}
\alias{compute_richness}
\title{Per-rank taxonomic richness from a long taxa tibble}
\usage{
compute_richness(taxa)
}
\description{
Per-rank taxonomic richness from a long taxa tibble
}
\keyword{internal}
