% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{count_seq_db}
\alias{count_seq_db}
\title{Count sequences in a FASTA file}
\usage{
count_seq_db(file)
}
\arguments{
\item{file}{(Character, required) Path to a FASTA file (plain or gzip).}
}
\value{
An integer, the number of sequences.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Counts the number of sequences in a FASTA file by counting header lines
(lines starting with \code{>}). Accepts gzip files.
}
\examples{
db <- system.file("extdata", "example_unite.fasta", package = "dbpq")
count_seq_db(db)
}
\seealso{
\code{\link[=count_pattern_db]{count_pattern_db()}}
}
\author{
Adrien Taudière
}
