% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{count_unwanted_in_headers}
\alias{count_unwanted_in_headers}
\title{Count unwanted taxonomic values within FASTA headers}
\usage{
count_unwanted_in_headers(headers)
}
\description{
Count unwanted taxonomic values within FASTA headers
}
\keyword{internal}
