% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpq-package.R
\docType{package}
\name{dbpq-package}
\alias{dbpq}
\alias{dbpq-package}
\title{dbpq: Manage FASTA Reference Databases for Taxonomic Assignment}
\description{
Download, format, summarize, and modify FASTA reference databases used for taxonomic assignment in metabarcoding pipelines. Supports major databases (UNITE, SILVA, PR2, BOLD, MaarjAM, Eukaryome) and converts between taxonomy header formats (dada2, SINTAX). Part of the 'pqverse' ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adrientaudiere/dbpq}
  \item Report bugs at \url{https://github.com/adrientaudiere/dbpq/issues}
}

}
\author{
\strong{Maintainer}: Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]

}
\keyword{internal}
