% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnosis_report}
\alias{diagnosis_report}
\title{Print a concise diagnosis report via message()}
\usage{
diagnosis_report(x)
}
\description{
Print a concise diagnosis report via message()
}
\keyword{internal}
