% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnosis_warning}
\alias{diagnosis_warning}
\title{Build a one-row warning tibble}
\usage{
diagnosis_warning(file, check, severity, message)
}
\description{
Build a one-row warning tibble
}
\keyword{internal}
