% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_unite_db}
\alias{download_unite_db}
\title{Download a UNITE reference database}
\usage{
download_unite_db(
  dest_dir = ".",
  type = c("dynamic", "static"),
  taxon_group = c("fungi", "eukaryotes"),
  verbose = TRUE
)
}
\arguments{
\item{dest_dir}{(Character, default \code{"."}) Directory to save the
downloaded file.}

\item{type}{(Character, default \code{"dynamic"}) One of \code{"dynamic"} or
\code{"static"}. Dynamic files include singletons, static do not.}

\item{taxon_group}{(Character, default \code{"fungi"}) One of \code{"fungi"} or
\code{"eukaryotes"}.}

\item{verbose}{(Logical, default \code{TRUE}) Print progress messages.}
}
\value{
The path to the downloaded file (invisibly).
}
\description{
Downloads the latest UNITE fungal ITS database for taxonomic assignment.
}
\examples{
\dontrun{
download_unite_db()
}
}
\author{
Adrien Taudière
}
