% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{extract_tax_matrix}
\alias{extract_tax_matrix}
\title{Extract a taxonomy matrix from a phyloseq object or FASTA file}
\usage{
extract_tax_matrix(x, tax_format = "auto")
}
\arguments{
\item{x}{A file path (character) or a phyloseq object.}

\item{tax_format}{Taxonomy format (only used for file input).}
}
\value{
A character matrix with rows = taxa and columns = ranks.
}
\description{
Extract a taxonomy matrix from a phyloseq object or FASTA file
}
\keyword{internal}
