% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{extract_taxa_long_db}
\alias{extract_taxa_long_db}
\title{Extract a long taxon table (file, rank, taxon, n_seqs) from a FASTA file}
\usage{
extract_taxa_long_db(file, tax_format = "auto")
}
\description{
Extract a long taxon table (file, rank, taxon, n_seqs) from a FASTA file
}
\keyword{internal}
