% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{filter_db}
\alias{filter_db}
\title{Filter a FASTA database by taxonomic pattern}
\usage{
filter_db(
  ref_fasta,
  pattern,
  output = NULL,
  force_two_lines_per_seq = TRUE,
  keep_temporary_files = FALSE
)
}
\arguments{
\item{ref_fasta}{(Character, required) Path to a FASTA file (plain or gzip).}

\item{pattern}{(Character, required) A pattern to search for in sequence
headers.}

\item{output}{(Character, required) Path to the output FASTA file
(must not be gzipped).}

\item{force_two_lines_per_seq}{(Logical, default \code{TRUE}) Force the FASTA
file to have exactly two lines per sequence (one header, one nucleotide
line). If \code{FALSE}, the input must already be in this format.}

\item{keep_temporary_files}{(Logical, default \code{FALSE}) If \code{TRUE} and
\code{force_two_lines_per_seq} is \code{TRUE}, keep intermediate temporary files.}
}
\value{
The path to the output file (invisibly).
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Filters sequences from a FASTA database whose header lines match a given
pattern. Accepts gzip files. May not work on Windows.
}
\examples{
\dontshow{if (tolower(Sys.info()[["sysname"]]) != "windows") withAutoprint(\{ # examplesIf}
db <- system.file("extdata", "example_unite.fasta", package = "dbpq")
out <- tempfile(fileext = ".fasta")
filter_db(db, "Amanita", output = out)
count_seq_db(out)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=count_pattern_db]{count_pattern_db()}}
}
\author{
Adrien Taudière
}
