% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get file extension(s)}
\usage{
get_file_extension(file_path)
}
\arguments{
\item{file_path}{(Character, required) Path to a file.}
}
\value{
A character vector of file extensions.
}
\description{
Get file extension(s)
}
\examples{
get_file_extension("my_database.fasta")
get_file_extension("my_database.fasta.gz")
}
