% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_gzipped}
\alias{is_gzipped}
\title{Check if a file is gzipped}
\usage{
is_gzipped(file_path)
}
\arguments{
\item{file_path}{(Character, required) Path to a file.}
}
\value{
Logical, TRUE if the file is gzipped.
}
\description{
Check if a file is gzipped
}
\keyword{internal}
