% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{is_vsearch_installed}
\alias{is_vsearch_installed}
\title{Check whether vsearch is installed}
\usage{
is_vsearch_installed(path = find_vsearch())
}
\arguments{
\item{path}{(Character, default \code{NULL}) Explicit path to the vsearch
executable. If \code{NULL}, the function searches the system PATH.}
}
\value{
\code{TRUE} if vsearch is available, \code{FALSE} otherwise.
}
\description{
Tests whether the \code{vsearch} executable is available on the system.
}
\examples{
is_vsearch_installed()
}
\seealso{
\code{\link[=find_vsearch]{find_vsearch()}}, \code{\link[=add_sh_to_taxonomy]{add_sh_to_taxonomy()}}
}
