% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_lines_db}
\alias{read_lines_db}
\title{Read lines from a plain or gzipped file}
\usage{
read_lines_db(file_path)
}
\arguments{
\item{file_path}{(Character, required) Path to a file (plain or gzip).}
}
\value{
A character vector of lines.
}
\description{
Read lines from a plain or gzipped file
}
\keyword{internal}
