% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{richness_plot}
\alias{richness_plot}
\title{Richness bar plot (distinct taxa per rank, grouped by file)}
\usage{
richness_plot(richness)
}
\description{
Richness bar plot (distinct taxa per rank, grouped by file)
}
\keyword{internal}
