% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{count_pattern_db}
\alias{count_pattern_db}
\title{Count lines matching a pattern in a FASTA file}
\usage{
count_pattern_db(file, pattern = ">")
}
\arguments{
\item{file}{(Character, required) Path to a file (plain or gzip),
often a FASTA file.}

\item{pattern}{(Character, default \code{">"}) A pattern to search for.}
}
\value{
An integer, the number of matching lines.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Count lines (sequences if fasta file) matching a pattern.
Accepts gzip files. May not work on Windows.
}
\examples{
\dontshow{if (tolower(Sys.info()[["sysname"]]) != "windows") withAutoprint(\{ # examplesIf}
db <- system.file("extdata", "example_unite.fasta", package = "dbpq")
count_pattern_db(db, "Amanita")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=filter_db]{filter_db()}}, \code{\link[=count_seq_db]{count_seq_db()}}
}
\author{
Adrien Taudière
}
