% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_tax_format}
\alias{detect_tax_format}
\title{Detect taxonomy format from FASTA headers}
\usage{
detect_tax_format(file, n_headers = 20L)
}
\arguments{
\item{file}{(Character, required) Path to a FASTA file (plain or gzip).}

\item{n_headers}{(Integer, default \code{20}) Number of headers to inspect.}
}
\value{
A character string: one of \code{"unite"}, \code{"sintax"},
\code{"greengenes2"}, \code{"pr2"}, or \code{"unknown"}.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Reads a few sequence headers from a FASTA file and guesses the
taxonomy format based on characteristic patterns.
}
\examples{
db <- system.file("extdata", "example_unite.fasta", package = "dbpq")
detect_tax_format(db)
}
\seealso{
\code{\link[=tax_prefixes]{tax_prefixes()}}, \code{\link[=list_ranks_db]{list_ranks_db()}}, \code{\link[=summarize_db]{summarize_db()}}
}
