% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_eukaryome_db}
\alias{download_eukaryome_db}
\title{Download the Eukaryome reference database}
\usage{
download_eukaryome_db(dest_dir = ".", verbose = TRUE)
}
\arguments{
\item{dest_dir}{(Character, default \code{"."}) Directory to save the
downloaded file.}

\item{verbose}{(Logical, default \code{TRUE}) Print progress messages.}
}
\value{
The path to the downloaded file (invisibly).
}
\description{
Downloads the Eukaryome database.
}
\examples{
\dontrun{
download_eukaryome_db()
}
}
\author{
Adrien Taudière
}
