% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_pr2_db}
\alias{download_pr2_db}
\title{Download a PR2 reference database}
\usage{
download_pr2_db(dest_dir = ".", version = NULL, verbose = TRUE)
}
\arguments{
\item{dest_dir}{(Character, default \code{"."}) Directory to save the
downloaded file.}

\item{version}{(Character) PR2 version number.}

\item{verbose}{(Logical, default \code{TRUE}) Print progress messages.}
}
\value{
The path to the downloaded file (invisibly).
}
\description{
Downloads the PR2 protist ribosomal reference database.
}
\examples{
\dontrun{
download_pr2_db()
}
}
\author{
Adrien Taudière
}
